Refactor to keep network logic isolated to corresponding utils.js
